Improved player output when using tpll command
